% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrdwabmisc_non-exported.R
\docType{data}
\name{CensusNames1990}
\alias{CensusNames1990}
\title{List of First Names and Surnames to Generate Random Names}
\format{A list of three \code{data.frame}s.
\describe{
  \item{surnames}{A \code{data.frame} of surnames and categories (common, 
  average, and rare).}
  \item{malenames}{A \code{data.frame} of male names and categories (common, 
  average, and rare).}
  \item{femalenames}{A \code{data.frame} of female names and categories 
  (common, average, and rare).}
}}
\source{
\emph{Genealogy Data: Frequently Occurring Surnames from Census
1990--Names Files}: \url{http://www.census.gov/genealogy/www/data/1990surnames/names_files.html}
}
\usage{
CensusNames1990
}
\description{
A list of first names and surnames by gender and category. This is the 
default dataset used by the \code{\link{RandomNames}} function.
}
\keyword{datasets}

