% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dfcols.list.R
\name{dfcols.list}
\alias{dfcols.list}
\title{Convert the Columns of a \code{data.frame} to a \code{list}}
\usage{
dfcols.list(data, vectorize = TRUE)
}
\arguments{
\item{data}{The input \code{data.frame}}

\item{vectorize}{Logical. Should the function return a list of single-column
\code{data.frame}s, or a simple \code{\link{vector}} of values? Defaults to
\code{TRUE}.}
}
\description{
Sometimes, it is useful to have the columns of a \code{data.frame} as
separate list items or vectors.  \code{\link{unlist}} is useful for creating
a single vector, but not for creating multiple vectors. The
\code{\link{dfcols.list}} function is a simple convenience function that
allows for such transformations.
}
\examples{

dat <- data.frame(A = c(1:2), B = c(3:4), C = c(5:6))
dfcols.list(dat)
dfcols.list(dat, vectorize = FALSE)
\dontshow{rm(dat)}

}
\author{
Ananda Mahto
}

