% Generated by roxygen2 (4.0.2): do not edit by hand
\name{SOQuestionPage}
\alias{SOQuestionPage}
\title{Download a Stack Overflow question page as a character vector}
\usage{
SOQuestionPage(qid)
}
\arguments{
\item{qid}{The numeric question ID.}
}
\value{
A character vector
}
\description{
Helper function for other Stack Overflow related functions. Downloads the
page as a character vector and extracts the portion that is needed for other
functions.
}
\author{
Ananda Mahto
}

