\docType{data}
\name{CensusNames1990}
\alias{CensusNames}
\alias{CensusNames1990}
\title{List of first names and surnames to generate random names}
\format{A list of first names and surnames, split by gender and how common
the first names are.}
\description{
  This is the default dataset used by the
  \code{\link{RandomNames}} function.
}
\references{
  \emph{Genealogy Data: Frequently Occurring Surnames from
  Census 1990--Names Files}:
  \url{http://www.census.gov/genealogy/www/data/1990surnames/names_files.html}
}
\keyword{datasets}

